Move enums files
