+++ psdm_analysis/models/grid_with_results.py
@@ -9,7 +9,7 @@ from psdm_analysis.models.input.container.grid import GridContainer
 from psdm_analysis.models.input.container.participants import (
     SystemParticipantsContainer,
 )
-from psdm_analysis.models.input.enums import RawGridElementsEnum
+from psdm_analysis.models.enums import RawGridElementsEnum
 from psdm_analysis.models.result.grid.connector import ConnectorsResult
 from psdm_analysis.models.result.grid.enhanced_node import EnhancedNodesResult
 from psdm_analysis.models.result.grid.node import NodesResult
