+++ psdm_analysis/models/input/connector/lines.py
@@ -5,7 +5,7 @@ from pandas import Series
 
 from psdm_analysis.models.input.connector.connector import Connector
 from psdm_analysis.models.input.container.mixins import HasTypeMixin
-from psdm_analysis.models.input.enums import RawGridElementsEnum
+from psdm_analysis.models.enums import RawGridElementsEnum
 
 
 @dataclass(frozen=True)
