+++ psdm_analysis/models/input/container/mixins.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 from psdm_analysis.io.utils import df_to_csv
 from psdm_analysis.models.input.connector.connector import Connector
 from psdm_analysis.models.input.entity import Entities
-from psdm_analysis.models.input.enums import EntitiesEnum
+from psdm_analysis.models.enums import EntitiesEnum
 from psdm_analysis.models.input.participant.participant import SystemParticipants
