+++ psdm_analysis/models/input/entity.py
@@ -13,7 +13,7 @@ from pandas import DataFrame, Series
 
 from psdm_analysis.io import utils
 from psdm_analysis.io.utils import bool_converter, df_to_csv, read_csv
-from psdm_analysis.models.input.enums import (
+from psdm_analysis.models.enums import (
     EntitiesEnum,
     RawGridElementsEnum,
     SystemParticipantsEnum,
