+++ psdm_analysis/models/input/node.py
@@ -6,7 +6,7 @@ from pandas import DataFrame
 
 from psdm_analysis.io.utils import df_to_csv
 from psdm_analysis.models.input.entity import Entities
-from psdm_analysis.models.input.enums import RawGridElementsEnum
+from psdm_analysis.models.enums import RawGridElementsEnum
 
 
 @dataclass(frozen=True)
