+++ psdm_analysis/models/input/participant/bm.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from psdm_analysis.models.input.container.mixins import SpTypeMixin
-from psdm_analysis.models.input.enums import SystemParticipantsEnum
+from psdm_analysis.models.enums import SystemParticipantsEnum
 from psdm_analysis.models.input.participant.participant import SystemParticipants
