+++ psdm_analysis/models/input/participant/evcs.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from psdm_analysis.models.input.enums import SystemParticipantsEnum
+from psdm_analysis.models.enums import SystemParticipantsEnum
 from psdm_analysis.models.input.participant.participant import SystemParticipants
