+++ psdm_analysis/models/input/participant/load.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import List
 
-from psdm_analysis.models.input.enums import SystemParticipantsEnum
+from psdm_analysis.models.enums import SystemParticipantsEnum
 from psdm_analysis.models.input.participant.participant import SystemParticipants
