+++ psdm_analysis/models/result/participant/dict.py
@@ -14,7 +14,7 @@ from psdm_analysis.io.utils import (
     to_date_time,
 )
 from psdm_analysis.models.input.entity import Entities
-from psdm_analysis.models.input.enums import EntitiesEnum, EntityEnumType
+from psdm_analysis.models.enums import EntitiesEnum, EntityEnumType
 from psdm_analysis.models.result.entity import ResultEntities
 
 ResultDictType = TypeVar("ResultDictType", bound="ResultDict")
