+++ psdm_analysis/models/result/participant/flex_options.py
@@ -6,7 +6,7 @@ from typing import Dict
 import pandas as pd
 from pandas import DataFrame
 
-from psdm_analysis.models.input.enums import SystemParticipantsEnum
+from psdm_analysis.models.enums import SystemParticipantsEnum
 from psdm_analysis.models.result.flex_option import FlexOptionResult
 from psdm_analysis.models.result.participant.dict import ResultDict
