+++ psdm_analysis/plots/common/line_plot.py
@@ -1,7 +1,7 @@
 from matplotlib.axes import Axes
 from pandas import Series
 
-from psdm_analysis.models.input.enums import EntitiesEnum
+from psdm_analysis.models.enums import EntitiesEnum
 from psdm_analysis.plots.common.utils import (
     FILL_ALPHA,
     add_to_kwargs_if_not_exist,
