+++ tests/models/input/container/test_grid.py
@@ -1,5 +1,5 @@
 from definitions import ROOT_DIR
-
+import os
 
 def test_grid_container(gwr):
     grid_container = gwr.grid
