+++ tests/models/input/test_enum.py
@@ -1,4 +1,4 @@
-from psdm_analysis.models.input.enums import RawGridElementsEnum, SystemParticipantsEnum
+from psdm_analysis.models.enums import RawGridElementsEnum, SystemParticipantsEnum
 
 
 def test_entities_enum():
