+++ tests/models/result/test_power.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 
 from psdm_analysis.io.utils import get_absolute_path
-from psdm_analysis.models.input.enums import SystemParticipantsEnum
+from psdm_analysis.models.enums import SystemParticipantsEnum
 from psdm_analysis.models.result.participant.participant import ParticipantsResult
 from psdm_analysis.models.result.power import PQResult
 from psdm_analysis.processing.series import duration_weighted_series
